Rename need-photo as do-you-have-photo
The functionality of a page is hard to depict if the name of the page
does not imply what it does.

"need-photo" answers the question "what was chosen" and requires reader
to know what the overall flow is in order to understand what the page
might contain.

"do-you-have-photo" answers the question "what does this page do" and
requires no external knowledge.